fix: remove dependency cycle in polymarket to avoid crashes
