+++ package.json
@@ -43,7 +43,7 @@
     "lint:js": "eslint --cache . --quiet",
     "lint:js:ci": "eslint --quiet",
     "lint:ts": "yarn tsc --skipLibCheck --noEmit",
-    "check:cycles": "./scripts/check-cycles.sh 2087",
+    "check:cycles": "./scripts/check-cycles.sh 2086",
     "postinstall": "./scripts/postinstall.sh",
     "start": "react-native start --client-logs",
     "start:clean": "yarn clean:packager && yarn start --reset-cache",
